Refactoring UserCnotroller

+++ src/main/java/com/ortizdavid/restapi/controllers/UserController.java
@@ -1,8 +1,6 @@
 package com.ortizdavid.restapi.controllers;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ortizdavid.restapi.entities.User;
 import com.ortizdavid.restapi.services.UserService;
+import com.ortizdavid.restapi.utils.Response;
 
 @RestController
 @RequestMapping("/api/users")
@@ -29,80 +28,52 @@ public class UserController {
     @GetMapping
     public ResponseEntity<?> getAllUsers() {
         List<User>users = userService.getAll();
-        Map<String, Object> response = new HashMap<String, Object>();
         if (users.size() == 0) {
-            response.put("status", HttpStatus.NOT_FOUND.value());
-            response.put("message", "No users found!");
-            response.put("count", 0);
-            response.put("data", null);
-            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+            return Response.build("Users not found!", null, 0, HttpStatus.NOT_FOUND);
         } else {
-            response.put("status", HttpStatus.OK.value());
-            response.put("message", "Users Found successfully!");
-            response.put("count", users.size());
-            response.put("data", users);
-            return new ResponseEntity<>(response, HttpStatus.OK);
+            return Response.build("Users found successfully!", users, users.size(), HttpStatus.OK);
         }
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getUser(@PathVariable Long id) {
         User user = userService.getById(id);
-        Map<String, Object> response = new HashMap<String, Object>();
         if (user == null) {
-            response.put("status", HttpStatus.NOT_FOUND.value());
-            response.put("message", "User not found!");
-            response.put("data", null);
-            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+            return Response.build("User not found!", null, 0, HttpStatus.NOT_FOUND);
         } else {
-            response.put("status", HttpStatus.OK.value());
-            response.put("message", "User found successfully!");
-            response.put("data", user);
-            return new ResponseEntity<>(response, HttpStatus.OK);
+            return Response.build("User found successfully!", user, 1, HttpStatus.OK);
         }
     }
 
     @PostMapping
     public ResponseEntity<?> createUser(@RequestBody User userBody) {
-        Map<String, Object> response = new HashMap<String, Object>();
         User user = userService.create(userBody);
         if (user == null) {
-            response.put("status", HttpStatus.BAD_REQUEST.value());
-            response.put("message", "Error while creating user!");
-            response.put("data", null);
-            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+            return Response.build("Error while creating user!", null, 0, HttpStatus.BAD_REQUEST);
         } else {
-            response.put("status", HttpStatus.CREATED.value());
-            response.put("message", "User created successfully!");
-            response.put("data", user);
-            return new ResponseEntity<>(response, HttpStatus.CREATED);
+            return Response.build("User created successfully!", user, 1, HttpStatus.CREATED);
         }
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<?> updateUser(@RequestBody User userBody, @PathVariable Long id) {
-        Map<String, Object> response = new HashMap<String, Object>();
         User user = userService.update(userBody, id);
         if (user == null) {
-            response.put("status", HttpStatus.BAD_REQUEST.value());
-            response.put("message", "Error while updating user!");
-            response.put("data", null);
-            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+            return Response.build("Error while updating user!", null, 0, HttpStatus.BAD_REQUEST);
         } else {
-            response.put("status", HttpStatus.CREATED.value());
-            response.put("message", "User updated successfully!");
-            response.put("data", user);
-            return new ResponseEntity<>(response, HttpStatus.CREATED);
+            return Response.build("User updated successfully!", user, 1, HttpStatus.OK);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable Long id) {
-        Map<String, Object> response = new HashMap<String, Object>();
-        userService.delete(id);
-        response.put("status", HttpStatus.OK.value());
-        response.put("message", "User deleted successfully!");
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+        User user = userService.getById(id);
+        if (user == null) {
+            return Response.build("Error while deleting user!", null, 0, HttpStatus.BAD_REQUEST);
+        } else {
+            userService.delete(id);
+            return Response.build("User deleted successfully!", null, 1, HttpStatus.OK);
+        }
     }
 
 }
